requirements.txt
cachelib==0.9.0
click==8.1.3
Flask==2.2.3
Flask-Caching==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
slack-sdk==3.20.2
Werkzeug==2.2.3
zipp==3.15.0
